Upon this first commit, we shall build our church
The baseline format for each page is set up, as well as the routing.
There is a navbar on every page that can bring you to either the
homescreen or views of your workouts, diets, and equipment. Each of
those pages displays the users previous entries of that content and has
an option to add an entry. That redirects to a new page, meaning there
is also an add view for diets, workouts, and equipment. I also added
meals within the diet section, but it isnt on the navbar so you have to
go through diets to get to that. If we want it on the navbar its ezpz so
we can decide that. Im gonna do some work on a different project now but
when I come back to this I think Imma start looking at user
authentication and setting that up. If you guys arent sure what to do,
either figuring out how our queries are going to look or maybe taking a
shot at setting up our initial database could help. If you guys know how
to pass data between views, we could also start trying to flesh out some
of those pages that I started.